updated Java files
