+++ javav2/usecases/workflow_multiple_channels/src/main/java/com/example/ListMissingStudentsHandler.java
@@ -1,0 +1,26 @@
+/*
+   Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+   SPDX-License-Identifier: Apache-2.0
+*/
+
+package com.example;
+
+import com.amazonaws.services.lambda.runtime.Context;
+import com.amazonaws.services.lambda.runtime.RequestHandler;
+import com.amazonaws.services.lambda.runtime.LambdaLogger;
+import java.util.Map;
+
+public class ListMissingStudentsHandler implements RequestHandler<Map<String,String>, String> {
+
+    @Override
+    public String handleRequest(Map<String,String> event, Context context) {
+        LambdaLogger logger = context.getLogger();
+        String date = event.get("date");
+        logger.log("DATE: " + date);
+
+        GetStudents students = new GetStudents();
+        String xml = students.getStudentsData(date);
+        logger.log("XML: " + xml);
+        return xml;
+    }
+}
