+++ javav2/usecases/workflow_multiple_channels/src/main/java/com/example/StudentData.java
@@ -1,0 +1,80 @@
+/*
+   Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+   SPDX-License-Identifier: Apache-2.0
+*/
+
+package com.example;
+
+import software.amazon.awssdk.enhanced.dynamodb.mapper.annotations.DynamoDbBean;
+import software.amazon.awssdk.enhanced.dynamodb.mapper.annotations.DynamoDbPartitionKey;
+import java.time.Instant;
+
+@DynamoDbBean
+public class StudentData {
+    private String id;
+
+    private String firstName;
+
+    private String lastName;
+    private String email;
+    private String mobileNumber ;
+    private String phoneNumber;
+
+    private Instant date;
+
+    public Instant getDate() {
+        return this.date;
+    }
+
+    public void setDate(Instant date) {
+        this.date = date;
+    }
+    @DynamoDbPartitionKey
+    public String getId() {
+        return this.id;
+    }
+
+    public void setId(String id) {
+      this.id = id;
+    }
+
+    public void setPhoneNunber(String phoneNunber) {
+        this.phoneNumber = phoneNunber;
+    }
+
+    public String getPhoneNunber() {
+        return this.phoneNumber;
+    }
+
+    public void setMobileNumber(String mobileNumber) {
+        this.mobileNumber = mobileNumber;
+    }
+
+    public String getMobileNumber() {
+        return this.mobileNumber;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getEmail() {
+        return this.email;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return this.lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public String getFirstName() {
+        return this.firstName;
+    }
+}
